refactor: 서버 상태 관련 로직을 server 디렉토리로 이동
