+++ src/components/ProductListPage/ProductList.tsx
@@ -8,7 +8,7 @@ import LoadingSpinner from "@components/common/LoadingSpinner";
 import { IntersectionDetector } from "@components/common/IntersectionDetector";
 
 import { useErrorToast } from "@contexts/errorToast/useErrorToast";
-import { useInfiniteProducts } from "@src/hooks/server/useInfiniteProducts";
+import { useInfiniteProducts } from "@server/useInfiniteProducts";
 
 const ProductList = () => {
   const { data, isLoading, error, fetchNextPage, updateCategoryFilter, updatePriceSort } =
