+++ src/server/useCartActions.ts
@@ -1,10 +1,9 @@
-import { CreateCartItemParams } from "@src/apis/cartItems";
-import { useCartItems } from "@hooks/server/useCartItems";
-import { OnError, useDeleteCartItemMutation } from "./useDeleteCartItemMutation";
-import { useCreateCartItemMutation } from "./useCreateCartItemMutation";
+import { useCartItems } from "@server/useCartItems";
+import { OnError, useDeleteCartItemMutation } from "@server/useDeleteCartItemMutation";
+import { useCreateCartItemMutation } from "@server/useCreateCartItemMutation";
 
 interface UseCartActionsReturn {
-  addToCart: (params: CreateCartItemParams) => Promise<void>;
+  addToCart: (productId: number) => void;
   removeFromCart: (productId: number) => void;
   isIncludedInCart: (productId: number) => boolean;
 }
@@ -13,19 +12,20 @@ const MAX_CART_ITEM_COUNT = 20;
 
 export const useCartActions = (onError: OnError = console.error): UseCartActionsReturn => {
   const { data: cartItems } = useCartItems();
+
   const { createCartItemMutation } = useCreateCartItemMutation(onError);
   const { deleteCartItemMutation } = useDeleteCartItemMutation(onError);
 
-  const addToCart = async (params: CreateCartItemParams) => {
+  const addToCart = async (productId: number) => {
     if (cartItems.length >= MAX_CART_ITEM_COUNT) {
       onError(Error("장바구니에 담을 수 있는 상품의 개수는 20개까지입니다."));
       return;
     }
 
-    createCartItemMutation(params);
+    createCartItemMutation({ productId, quantity: 1 });
   };
 
-  const removeFromCart = (productId: number) => {
+  const removeFromCart = async (productId: number) => {
     const targetCartItemId = cartItems.find((cartItem) => cartItem.product.id === productId)?.id;
 
     if (targetCartItemId === undefined) {
