+++ src/server/useDeleteCartItemMutation.ts
@@ -1,6 +1,7 @@
 import { deleteCartItem } from "@src/apis/cartItems";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { MUTATION_KEYS, QUERY_KEYS } from "../__constants__/queryKeys";
+import { useMutation } from "@tanstack/react-query";
+import { MUTATION_KEYS, QUERY_KEYS } from "./__constants__/queryKeys";
+import { queryClient } from "./queryClient";
 
 interface UseDeleteCartItemMutationReturn {
   deleteCartItemMutation: (cartItemId: number) => void;
@@ -9,14 +10,14 @@ interface UseDeleteCartItemMutationReturn {
 export type OnError = (error: Error) => void;
 
 export const useDeleteCartItemMutation = (onError?: OnError): UseDeleteCartItemMutationReturn => {
-  const queryClient = useQueryClient();
+  const invalidateCartItems = () => {
+    queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.cartItems] });
+  };
 
   const { mutate } = useMutation({
-    mutationKey: [MUTATION_KEYS.removeCartItem],
+    mutationKey: [MUTATION_KEYS.deleteCartItem],
     mutationFn: deleteCartItem,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.cartItems] });
-    },
+    onSuccess: invalidateCartItems,
     onError,
   });
